feat(plugins): enhance plugin configurations
- easypick plugin now previews changes against last 5 commits
- lsp-trouble plugin initialization is no longer lazy
- vim-ai plugin updated to use gpt-4 model by default

The easypick plugin configuration was updated to include a more
informative previewer that shows changes in the working tree against
the last 5 commits, to help identify recent file changes more easily.

Lsp-trouble plugin's 'lazy' option was set to false to ensure that the
plugin loads its dependencies immediately on startup.

Vim-ai plugin's configuration was enhanced with a dictionary of models.
It now defaults to the gpt-4-1106-preview model, reflecting the latest
advancements in AI technology for more fluid coding assistance.