feat: update plugins and add new ones
Update `lazy-lock.json` with new plugin commits and add entries for
`boo.nvim`, `fzf-lua`, `harpoon`, and others.

Also, extend the `lua/commands.lua` with a new `HarpoonHelp` command that
displays a help menu for Harpoon in a floating window.

New plugins and breaking changes to the `lazy-plugins.lua` structure are:
1. vim-sleuth
2. boo.nvim
3. fzf-lua
4. harpoon
5. highlight-undo.nvim
6. hoversplit.nvim
7. nvim-treesitter-context
8. oil.nvim
9. ollama.nvim
10. sg.nvim
11. telescope-egrepify.nvim
12. tsc.nvim
13. watch.nvim
14. git-blame.nvim